requirements.txt
schedule
kafka-python
cassandra-driver